Bug fix: Used wrong data type extension for variable m

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -58,7 +58,7 @@ int clear_bit(unsigned long int *n, unsigned int index)
  *
  * Return: number of bits
  **/
-int flip_bits(unsigned long int n, unsigned int m)
+int flip_bits(unsigned long int n, unsigned long int m)
 {
 	int bits = 0, index;
